refactor(vault): move merge_frontmatter to frontmatter.py
backfill_frontmatter.py (a core CLI) already reached into the skill package
for it, and the vault REST API needs it next — three call sites. Beside the
parser it also loses its cycle-breaking function-level import.
